Example code to get the API version of VTN Manager.
This commit shows how to get the API verions of VTN Manager. To get the
API version, the example application calls the get-vtn-version RPC
provided by the VTN Manager.

The example application gets the API version during its initialization,
and logs it. You can check the log message which shows the result of the
execution of the get-vtn-version RPC by executing the following command
in the Karaf console.

    opendaylight-user@root>log:display | grep VTN